fix: split intelligence digest models under 300-line cap (#4476)
